+++ packages/core/src/hooks/useBlockMeta.test.tsx
@@ -14,7 +14,7 @@ describe('useBlockMeta', () => {
   })
   it('retrieves block timestamp and difficulty', async () => {
     const { result, waitForCurrent } = await renderDAppHook(useBlockMeta, { config })
-    await waitForCurrent((val) => val.timestamp !== undefined && val.difficulty !== undefined)
+    await waitForCurrent((val) => val?.timestamp !== undefined && val?.difficulty !== undefined)
 
     expect(result.error).to.be.undefined
     expect(result.current.timestamp).to.be.a('date')
@@ -29,8 +29,7 @@ describe('useBlockMeta', () => {
     const firstTimestamp = result.current.timestamp
 
     await sleep(1000)
-    await network1.wallets[0].sendTransaction({ to: receiver, value: 100 })
-    await sleep(1000)
+    await network1.mineBlock()
     await waitForCurrent((val) => val.timestamp?.getTime() !== firstTimestamp?.getTime())
     expect(result.current.timestamp).to.be.greaterThan(firstTimestamp)
   })
@@ -47,5 +46,31 @@ describe('useBlockMeta', () => {
     await sleep(1000)
     await waitForCurrent((val) => val.timestamp?.getTime() !== firstTimestamp?.getTime())
     expect(result.current.timestamp).to.be.greaterThan(firstTimestamp)
+  })
+
+  it('updates the block number when a transaction gets mined', async () => {
+    const { config, network1 } = await setupTestingConfig()
+    const { result, waitForCurrent } = await renderDAppHook(useBlockMeta, { config })
+    const blockNumberFromProvider = await network1.provider.getBlockNumber()
+    await waitForCurrent(({ blockNumber }) => blockNumber === blockNumberFromProvider)
+
+    await network1.mineBlock()
+
+    await waitForCurrent(({ blockNumber }) => blockNumber === blockNumberFromProvider + 1)
+    expect(result.error).to.be.undefined
+  })
+
+  it('updates the block number when a transaction gets mined on another chain', async () => {
+    const { config, network2 } = await setupTestingConfig()
+    const { result, waitForCurrent } = await renderDAppHook(() => useBlockMeta({ chainId: network2.chainId }), {
+      config,
+    })
+    const blockNumberFromProvider = await network2.provider.getBlockNumber()
+    await waitForCurrent(({ blockNumber }) => blockNumber === blockNumberFromProvider)
+
+    await network2.mineBlock()
+
+    await waitForCurrent(({ blockNumber }) => blockNumber === blockNumberFromProvider + 1)
+    expect(result.error).to.be.undefined
   })
 })