+++ packages/testing/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @usedapp/testing
 
+## 1.1.3
+
+### Patch Changes
+
+- Updated dependencies [f2c3766]
+- Updated dependencies [3becd1d]
+  - @usedapp/core@1.1.3
+
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies [bbb0cc9]
+  - @usedapp/core@1.1.2
+
 ## 1.1.1
 
 ### Patch Changes